* 3035_cleanup: (29 commits)
  Indentation.
  doc/NEWS: cosmetics.
  Typo fixes.
  Fix panel recoding
  (edit_status_window): type accuracy in status displaying.
  (edit_get_save_file_as): remove extra code block.
  (symlink_dialog): indentation cosmetics.
  (panel_change_encoding): use vfs_path_as_str().
  Fix for opening some jar files
  Clarify midnight_dlg size in MSG_RESIZE handling.
  Ticket #3043: broken NCurses detection.
  Adjust progress window for file delete operation.
  (move_dir_dir): get rid of extra memory allocation.
  (copy_dir_dir): get rid of extra memory allocation.
  (erase_dir_iff_empty): get rid of extra directory name conversion.
  (recursive_erase): ret rid of extra path conversion.
  (my_mkdir): don't call mc_mkdir() twice in case of failure.
  (my_mkdir_rec): ret rid of double directory name conversion.
  (erase_dir_iff_empty): optimization.
  Add DIR_IS_DOT and DIR_IS_DOTDOT macros
  ...

+++ src/editor/editdraw.c
@@ -294,8 +294,8 @@ edit_status_window (WEdit * edit)
         edit_move (2, w->lines - 1);
         tty_printf ("%3ld %5ld/%ld %6ld/%ld",
                     edit->curs_col + edit->over_col,
-                    edit->buffer.curs_line + 1, edit->buffer.lines + 1, edit->buffer.curs1,
-                    edit->buffer.size);
+                    edit->buffer.curs_line + 1, edit->buffer.lines + 1, (long) edit->buffer.curs1,
+                    (long) edit->buffer.size);
     }
 
     /*
